(all) re-arrange actions in AbstractAppender.internalAppend()
There were two problems: (1) a reasonable assumption that writer would be not-null
on entry, and (2) a rather unclear test for a null writer before writing messages,
driven by the possibility of rotate() failing. Moved the rotation test to the top
of the synchronized block, and put the null-after-rotate test inside the rotation
block.